+++ saffier/db/constants.py
@@ -3,7 +3,6 @@ ORDER_DIR = {
     "ASC": ("ASC", "DESC"),
     "DESC": ("DESC", "ASC"),
 }
-MAX_GET_RESULTS = 21
 
 FILTER_OPERATORS = {
     "exact": "__eq__",
@@ -22,6 +21,4 @@ DEFAULT_RELATED_LOOKUP_FIELD = "id"
 CASCADE = "CASCADE"
 RESTRICT = "RESTRICT"
 SET_NULL = "SET NULL"
-REPR_OUTPUT_SIZE = 20
-SAFFIER_PICKLE_KEY = "saffier-version"
 MANY_TO_MANY_RELATION = "relation_{key}"
